refactor: code file breakdown and reorg
 * Break actions into multiple files under /lib/actions
 * Break decorators into multiple files under /lib/decorators
 * Break effects into multiple files under /lib/effects
 * Move operators under /lib/effects
 * Move reducer under /lib/reducer
 * Break service down into multiple files under /lib/service
 * Break util down into multiple files under /lib/util
 * Update library index to properly reference all files
 * Update each area of library to properly reference files
 * Move uuid generator function to /util
