+++ projects/ngrx-auto-entity/src/index.ts
@@ -3,20 +3,6 @@
  */
 
 export {
-  buildFeatureState,
-  buildState,
-  buildSelectorMap,
-  buildFacade,
-  IEntityDictionary,
-  IEntityState,
-  IEntityFacade,
-  IModelState,
-  ISelectorMap,
-  IModelClass,
-  EntityIdentity
-} from './lib/util';
-
-export {
   IPage,
   Page,
   IFirstLastRange,
@@ -28,16 +14,23 @@ export {
   IPageInfo
 } from './lib/models';
 
+export { IEntityDictionary, IEntityState, EntityIdentity } from './lib/util/entity-state';
+export { IModelState, IModelClass } from './lib/util/model-state';
+export { IEntityFacade } from './lib/util/facade';
+export { ISelectorMap } from './lib/util/selector-map';
+export { buildFacade } from './lib/util/facade-builder';
+export { buildSelectorMap } from './lib/util/selector-map-builder';
+export { buildFeatureState, buildState } from './lib/util/state-builder';
+
+export { EntityActionTypes } from './lib/actions/action-types';
+export { IEntityInfo } from './lib/actions/entity-info';
+export { EntityAction } from './lib/actions/entity-action';
+export { ICorrelatedAction } from './lib/actions/entity-action';
+export { fromEntityActions } from './lib/actions/action-operators';
+export { ofEntityType } from './lib/actions/action-operators';
+export { ofEntityAction } from './lib/actions/action-operators';
+export { isEntityActionInstance, EntityActions } from './lib/actions/entity-actions-union';
 export {
-  ICorrelatedAction,
-  ofEntityType,
-  ofEntityAction,
-  fromEntityActions,
-  isEntityActionInstance,
-  IEntityInfo,
-  EntityActions,
-  EntityAction,
-  EntityActionTypes,
   Load,
   LoadFailure,
   LoadSuccess,
@@ -103,62 +96,52 @@ export {
   EndEdit,
   EditEnded,
   Clear
-} from './lib/actions';
+} from './lib/actions/actions';
 
-export {
-  IEffectExclusions,
-  IEffectExcept,
-  IEntityOptions,
-  IEntityTransformer,
-  Entity,
-  all,
-  extra,
-  loads,
-  curd,
-  except,
-  matching
-} from './lib/decorators/entity';
-
+export { ENTITY_OPTS_PROP } from './lib/decorators/entity-tokens';
+export { IEffectExcept, IEntityOptions, IEntityTransformer, Entity } from './lib/decorators/entity';
+export { IEffectExclusions } from './lib/decorators/effect-exclusions';
+export { curd, loads, extra, all, matching, except } from './lib/decorators/effect-exclusion-utils';
 export { Key, getKey, getKeyFromModel, getKeyNames, getKeyNamesFromModel, checkKeyName } from './lib/decorators/key';
 
-export { EntityOperators } from './lib/operators';
+export { EntityOperators } from './lib/effects/operators';
 
-export { autoEntityReducer, autoEntityMetaReducer, stateNameFromAction } from './lib/reducer';
+export { autoEntityReducer, autoEntityMetaReducer, stateNameFromAction } from './lib/reducer/reducer';
 
 export {
-  NgrxAutoEntityService,
-  IEntityWithRangeInfo,
-  IEntityWithPageInfo,
   IEntityRangeRef,
   IEntityPageRef,
-  IEntityError,
-  IAutoEntityService,
   IEntityRef,
   IEntityIdentityRef,
   IEntityIdentitiesRef
-} from './lib/service';
+} from './lib/service/refs';
+export { IEntityWithRangeInfo, IEntityWithPageInfo, IEntityError } from './lib/service/wrapper-models';
+export { IAutoEntityService } from './lib/service/interface';
+export { NgrxAutoEntityService } from './lib/service/service';
 
+export { EntityEffects } from './lib/effects/effects-all';
+export { LoadEffects } from './lib/effects/effects-loads';
+export { CUDEffects } from './lib/effects/effects-cud';
+export { ExtraEffects } from './lib/effects/effects-extra';
 export {
-  LoadEffects,
   LoadEffect,
   LoadAllEffect,
   LoadManyEffect,
   LoadPageEffect,
-  LoadRangeEffect,
-  CUDEffects,
+  LoadRangeEffect
+} from './lib/effects/effects-loads-discrete';
+export {
   CreateEffect,
   CreateManyEffect,
-  UpdateEffect,
-  UpdateManyEffect,
+  DeleteEffect,
+  DeleteManyEffect,
+  DeleteByKeyEffect,
+  DeleteManyByKeysEffect,
   ReplaceEffect,
   ReplaceManyEffect,
-  DeleteEffect,
-  DeleteByKeyEffect,
-  DeleteManyEffect,
-  DeleteManyByKeysEffect,
-  EntityEffects,
-  ExtraEffects
-} from './lib/effects.default';
+  UpdateEffect,
+  UpdateManyEffect
+} from './lib/effects/effects-cud-discrete';
 
 export {
   NgrxAutoEntityModule,
