+++ projects/ngrx-auto-entity/src/lib/actions/actions.ts
@@ -1,172 +1,7 @@
-import { Actions } from '@ngrx/effects';
-import { Action } from '@ngrx/store';
-import { merge, Observable, OperatorFunction } from 'rxjs';
-import { filter } from 'rxjs/operators';
-
-import { pascalCase } from '../util/case';
-import { IEntityOptions, IEntityTransformer } from './decorators/entity';
-import { ENTITY_OPTS_PROP } from './decorators/entity-tokens';
-import { checkKeyName } from './decorators/key';
-import { IPageInfo, IRangeInfo, Page, Range } from './models';
-import { EntityIdentity } from './util';
-
-export enum EntityActionTypes {
-  Load = '[Entity] (Generic) Load',
-  LoadSuccess = '[Entity] (Generic) Load: Success',
-  LoadFailure = '[Entity] (Generic) Load: Failure',
-
-  LoadMany = '[Entity] (Generic) Load Many',
-  LoadManySuccess = '[Entity] (Generic) Load Many: Success',
-  LoadManyFailure = '[Entity] (Generic) Load Many: Failure',
-
-  LoadAll = '[Entity] (Generic) Load All',
-  LoadAllSuccess = '[Entity] (Generic) Load All: Success',
-  LoadAllFailure = '[Entity] (Generic) Load All: Failure',
-
-  LoadPage = '[Entity] (Generic) Load Page',
-  LoadPageSuccess = '[Entity] (Generic) Load Page: Success',
-  LoadPageFailure = '[Entity] (Generic) Load Page: Failure',
-
-  LoadRange = '[Entity] (Generic) Load Range',
-  LoadRangeSuccess = '[Entity] (Generic) Load Range: Success',
-  LoadRangeFailure = '[Entity] (Generic) Load Range: Failure',
-
-  Create = '[Entity] (Generic) Create',
-  CreateSuccess = '[Entity] (Generic) Create: Success',
-  CreateFailure = '[Entity] (Generic) Create: Failure',
-
-  CreateMany = '[Entity] (Generic) Create Many',
-  CreateManySuccess = '[Entity] (Generic) Create Many: Success',
-  CreateManyFailure = '[Entity] (Generic) Create Many: Failure',
-
-  Update = '[Entity] (Generic) Update',
-  UpdateSuccess = '[Entity] (Generic) Update: Success',
-  UpdateFailure = '[Entity] (Generic) Update: Failure',
-
-  UpdateMany = '[Entity] (Generic) Update Many',
-  UpdateManySuccess = '[Entity] (Generic) Update Many: Success',
-  UpdateManyFailure = '[Entity] (Generic) Update Many: Failure',
-
-  Replace = '[Entity] (Generic) Replace',
-  ReplaceSuccess = '[Entity] (Generic) Replace: Success',
-  ReplaceFailure = '[Entity] (Generic) Replace: Failure',
-
-  ReplaceMany = '[Entity] (Generic) Replace Many',
-  ReplaceManySuccess = '[Entity] (Generic) Replace Many: Success',
-  ReplaceManyFailure = '[Entity] (Generic) Replace Many: Failure',
-
-  Delete = '[Entity] (Generic) Delete',
-  DeleteSuccess = '[Entity] (Generic) Delete: Success',
-  DeleteFailure = '[Entity] (Generic) Delete: Failure',
-
-  DeleteMany = '[Entity] (Generic) Delete Many',
-  DeleteManySuccess = '[Entity] (Generic) Delete Many: Success',
-  DeleteManyFailure = '[Entity] (Generic) Delete Many: Failure',
-
-  DeleteByKey = '[Entity] (Generic) Delete by key',
-  DeleteByKeySuccess = '[Entity] (Generic) Delete by key: Success',
-  DeleteByKeyFailure = '[Entity] (Generic) Delete by key: Failure',
-
-  DeleteManyByKeys = '[Entity] (Generic) Delete many by keys',
-  DeleteManyByKeysSuccess = '[Entity] (Generic) Delete many by keys: Success',
-  DeleteManyByKeysFailure = '[Entity] (Generic) Delete many by keys: Failure',
-
-  Clear = '[Entity] (Generic) Clear',
-
-  Select = '[Entity] (Generic) Select',
-  SelectByKey = '[Entity] (Generic) Select by Key',
-  Selected = '[Entity] (Generic) Selection',
-
-  SelectMany = '[Entity] (Generic) Select Many',
-  SelectMore = '[Entity] (Generic) Select More',
-  SelectManyByKeys = '[Entity] (Generic) Select Many by Keys',
-  SelectMoreByKeys = '[Entity] (Generic) Select More by Keys',
-  SelectedMany = '[Entity] (Generic) Selection of Many',
-  SelectedMore = '[Entity] (Generic) Selection of More',
-
-  Deselect = '[Entity] (Generic) Deselect',
-  Deselected = '[Entity] (Generic) Deselection',
-
-  DeselectMany = '[Entity] (Generic) Deselect of Many',
-  DeselectManyByKeys = '[Entity] (Generic) Deselect of Many by Keys',
-  DeselectAll = '[Entity] (Generic) Deselect of All',
-  DeselectedMany = '[Entity] (Generic) Deselection of Many',
-
-  Edit = '[Entity] (Generic) Edit',
-  Edited = '[Entity] (Generic) Edited',
-  Change = '[Entity] (Generic) Change',
-  Changed = '[Entity] (Generic) Changed',
-  EndEdit = '[Entity] (Generic) Edit: End',
-  EditEnded = '[Entity] (Generic) Edit: Ended'
-}
-
-/**
- * Entity class
- */
-export interface IEntityInfo {
-  modelName: string;
-  pluralName?: string;
-  uriName?: string;
-  modelType: new () => any;
-  transform?: IEntityTransformer[];
-}
-
-export type TNew<TModel> = new () => TModel;
-
-export interface ICorrelatedAction {
-  correlationId: string;
-}
-
-export interface IEntityAction extends Action, ICorrelatedAction {
-  actionType: string;
-  info: IEntityInfo;
-}
-
-// tslint:disable
-const uuid = (a?, b?) => {
-  for (
-    b = a = '';
-    a++ < 36;
-    b += (a * 51) & 52 ? (a ^ 15 ? 8 ^ (Math.random() * (a ^ 20 ? 16 : 4)) : 4).toString(16) : '-'
-  );
-  return b;
-};
-
-// tslint:enable
-
-/**
- * Structure for all of this library's actions
- */
-export abstract class EntityAction<TModel> implements IEntityAction {
-  type: string;
-  info: IEntityInfo;
-
-  protected constructor(type: TNew<TModel>, public actionType: string, public correlationId: string = uuid()) {
-    this.info = setInfo(type);
-    this.type = setType(this.actionType, this.info);
-  }
-}
-
-const setInfo = (type: any): IEntityInfo => {
-  const instance = new type();
-  const opts = (type[ENTITY_OPTS_PROP] || { modelName: instance.constructor.name }) as IEntityOptions;
-  const modelName = opts.modelName;
-  checkKeyName(type, modelName);
-  return {
-    modelType: type,
-    modelName,
-    pluralName: opts.pluralName,
-    uriName: opts.uriName,
-    transform: opts.transform
-  };
-};
-
-const setType = (actionType: string, info: IEntityInfo): string => {
-  const name = info.modelName;
-  const entity = pascalCase(name);
-
-  return actionType.replace('Entity', entity);
-};
+import { IPageInfo, IRangeInfo, Page, Range } from '../models';
+import { EntityIdentity } from '../util/entity-state';
+import { EntityActionTypes } from './action-types';
+import { EntityAction } from './entity-action';
 
 /**
  * Loads a single instance of an entity, corresponding to HTTP GET /entity/:id operation
@@ -731,181 +566,4 @@ export class EditEnded<TModel> extends EntityAction<TModel> {
   constructor(type: new () => TModel, correlationId?: string) {
     super(type, EntityActionTypes.EditEnded, correlationId);
   }
-}
-
-/**
- * Union of all known entity action types
- */
-export type EntityActions<TModel> =
-  | Load<TModel>
-  | LoadFailure<TModel>
-  | LoadSuccess<TModel>
-  | LoadMany<TModel>
-  | LoadManyFailure<TModel>
-  | LoadManySuccess<TModel>
-  | LoadAll<TModel>
-  | LoadAllFailure<TModel>
-  | LoadAllSuccess<TModel>
-  | LoadPage<TModel>
-  | LoadPageFailure<TModel>
-  | LoadPageSuccess<TModel>
-  | LoadRange<TModel>
-  | LoadRangeFailure<TModel>
-  | LoadRangeSuccess<TModel>
-  | Create<TModel>
-  | CreateFailure<TModel>
-  | CreateSuccess<TModel>
-  | CreateMany<TModel>
-  | CreateManyFailure<TModel>
-  | CreateManySuccess<TModel>
-  | Update<TModel>
-  | UpdateFailure<TModel>
-  | UpdateSuccess<TModel>
-  | UpdateMany<TModel>
-  | UpdateManyFailure<TModel>
-  | UpdateManySuccess<TModel>
-  | Replace<TModel>
-  | ReplaceFailure<TModel>
-  | ReplaceSuccess<TModel>
-  | ReplaceMany<TModel>
-  | ReplaceManyFailure<TModel>
-  | ReplaceManySuccess<TModel>
-  | Delete<TModel>
-  | DeleteFailure<TModel>
-  | DeleteSuccess<TModel>
-  | DeleteMany<TModel>
-  | DeleteManyFailure<TModel>
-  | DeleteManySuccess<TModel>
-  | DeleteByKey<TModel>
-  | DeleteByKeyFailure<TModel>
-  | DeleteByKeySuccess<TModel>
-  | DeleteManyByKeys<TModel>
-  | DeleteManyByKeysFailure<TModel>
-  | DeleteManyByKeysSuccess<TModel>
-  | Clear<TModel>
-  | Select<TModel>
-  | SelectByKey<TModel>
-  | SelectMany<TModel>
-  | SelectManyByKeys<TModel>
-  | Selected<TModel>
-  | SelectedMany<TModel>
-  | Deselect<TModel>
-  | DeselectMany<TModel>
-  | DeselectManyByKeys<TModel>
-  | Deselected<TModel>
-  | DeselectedMany<TModel>
-  | Edit<TModel>
-  | Edited<TModel>
-  | Change<TModel>
-  | Changed<TModel>
-  | EndEdit<TModel>
-  | EditEnded<TModel>;
-
-export const isEntityActionInstance = (action: IEntityAction): boolean =>
-  action instanceof Load ||
-  action instanceof LoadSuccess ||
-  action instanceof LoadFailure ||
-  action instanceof LoadMany ||
-  action instanceof LoadManySuccess ||
-  action instanceof LoadManyFailure ||
-  action instanceof LoadAll ||
-  action instanceof LoadAllSuccess ||
-  action instanceof LoadAllFailure ||
-  action instanceof LoadPage ||
-  action instanceof LoadPageSuccess ||
-  action instanceof LoadPageFailure ||
-  action instanceof LoadRange ||
-  action instanceof LoadRangeSuccess ||
-  action instanceof LoadRangeFailure ||
-  action instanceof Create ||
-  action instanceof CreateSuccess ||
-  action instanceof CreateFailure ||
-  action instanceof CreateMany ||
-  action instanceof CreateManySuccess ||
-  action instanceof CreateManyFailure ||
-  action instanceof Update ||
-  action instanceof UpdateSuccess ||
-  action instanceof UpdateFailure ||
-  action instanceof UpdateMany ||
-  action instanceof UpdateManySuccess ||
-  action instanceof UpdateManyFailure ||
-  action instanceof Replace ||
-  action instanceof ReplaceSuccess ||
-  action instanceof ReplaceFailure ||
-  action instanceof ReplaceMany ||
-  action instanceof ReplaceManySuccess ||
-  action instanceof ReplaceManyFailure ||
-  action instanceof Delete ||
-  action instanceof DeleteSuccess ||
-  action instanceof DeleteFailure ||
-  action instanceof DeleteMany ||
-  action instanceof DeleteManySuccess ||
-  action instanceof DeleteManyFailure ||
-  action instanceof DeleteByKey ||
-  action instanceof DeleteByKeySuccess ||
-  action instanceof DeleteByKeyFailure ||
-  action instanceof DeleteManyByKeys ||
-  action instanceof DeleteManyByKeysSuccess ||
-  action instanceof DeleteManyByKeysFailure ||
-  action instanceof Clear ||
-  action instanceof Select ||
-  action instanceof SelectByKey ||
-  action instanceof SelectMany ||
-  action instanceof SelectManyByKeys ||
-  action instanceof Selected ||
-  action instanceof SelectedMany ||
-  action instanceof Deselect ||
-  action instanceof DeselectMany ||
-  action instanceof DeselectManyByKeys ||
-  action instanceof DeselectAll ||
-  action instanceof Deselected ||
-  action instanceof DeselectedMany ||
-  action instanceof Edit ||
-  action instanceof Edited ||
-  action instanceof Change ||
-  action instanceof Changed ||
-  action instanceof EndEdit ||
-  action instanceof EditEnded;
-
-/**
- * Operator to filter actions by an entity action type or multiple action types.
- *
- * @param allowedActionTypes One or more action type string constants
- */
-export function ofEntityAction<T extends IEntityAction>(
-  ...allowedActionTypes: EntityActionTypes[]
-): OperatorFunction<Action, T> {
-  return filter((action: IEntityAction): action is T => {
-    return isEntityActionInstance(action)
-      ? allowedActionTypes.some(type => setType(type, action.info) === action.type)
-      : false;
-  });
-}
-
-/**
- * Operator to filter actions by an entity and action type or multiple action types.
- *
- * @param entity The entity class
- * @param allowedActionTypes One or more action type string constants
- */
-export function ofEntityType<TModel, T extends EntityAction<TModel>>(
-  entity: new () => TModel,
-  ...allowedActionTypes: EntityActionTypes[]
-): OperatorFunction<Action, T> {
-  return filter((action: EntityAction<TModel>): action is T => {
-    // console.log('ofEntityType action', JSON.stringify(action, undefined, 2), 'setType');
-    // allowedActionTypes.forEach(type => console.log('setType', setType(type, action.info), '===', action.type));
-    return isEntityActionInstance(action)
-      ? action.info.modelType === entity && allowedActionTypes.some(type => setType(type, action.info) === action.type)
-      : false;
-  });
-}
-
-export function fromEntityActions<T extends EntityAction<any>>(
-  actions$: Actions,
-  entity: Array<new () => any>,
-  ...allowedActionTypes: EntityActionTypes[]
-): Observable<Action> {
-  const entityActions = entity.map(e => actions$.pipe(ofEntityType(e, ...allowedActionTypes)));
-  return merge(...entityActions);
 }
