+++ projects/ngrx-auto-entity/src/lib/actions/model-constructor.ts
@@ -1,0 +1,4 @@
+/**
+ * Constructor type for an Entity model.
+ */
+export type TNew<TModel> = new () => TModel;
