--- projects/ngrx-auto-entity/src/lib/index.ts
@@ -1,8 +1,0 @@
-export * from './util';
-export * from './actions';
-export * from './decorators/entity';
-export * from './decorators/entity-constants';
-export * from './decorators/key';
-export * from './module';
-export * from './reducer';
-export * from './models';
