+++ projects/ngrx-auto-entity/src/lib/service/service-resolution.ts
@@ -1,0 +1,37 @@
+import { Injector } from '@angular/core';
+
+import { pascalCase } from '../../util/case';
+import { IEntityInfo } from '../actions/entity-info';
+import { logErrorDetails, logServiceLocateFailure } from './error-handling';
+import { IAutoEntityService } from './interface';
+
+export const failResolution = (error: any, entityInfo: IEntityInfo): void => {
+  const serviceName = `${pascalCase(entityInfo.modelName)}Service`;
+  logServiceLocateFailure(entityInfo, serviceName);
+  logErrorDetails(error);
+  throw error;
+};
+
+export const resolveService = <TModel>(
+  entityInfo: Readonly<IEntityInfo>,
+  injector: Injector
+): IAutoEntityService<TModel> => {
+  return injector.get(entityInfo.modelType);
+};
+
+export const resolveServiceDeep = <TModel>(
+  entityInfo: Readonly<IEntityInfo>,
+  injector: Injector,
+  remaining: Injector[]
+): IAutoEntityService<TModel> => {
+  try {
+    return resolveService(entityInfo, injector);
+  } catch (err) {
+    if (remaining.length) {
+      const [first, ...rest] = remaining;
+      return resolveServiceDeep(entityInfo, first, rest);
+    } else {
+      failResolution(err, entityInfo);
+    }
+  }
+};
