+++ projects/ngrx-auto-entity/src/lib/service/wrapper-models.ts
@@ -1,0 +1,18 @@
+import { IEntityInfo } from '../actions/entity-info';
+import { IPageInfo, IRangeInfo } from '../models';
+
+export interface IEntityError<TModel> {
+  info: IEntityInfo;
+  message?: string;
+  err?: any;
+}
+
+export interface IEntityWithPageInfo<TModel> {
+  entities: TModel[];
+  pageInfo: IPageInfo;
+}
+
+export interface IEntityWithRangeInfo<TModel> {
+  entities: TModel[];
+  rangeInfo: IRangeInfo;
+}
