+++ projects/ngrx-auto-entity/src/lib/util/util-tokens.ts
@@ -1,0 +1,1 @@
+export const FEATURE_AFFINITY = '__ngrxae_feature_affinity';
